Added script
Man I spent quite a time on this script. Yet it's not completely
finished.
However it does parse the ADE url providing the default login and
password.
The all caps variables can be manipulated to get the desired result.
As time of commit, it fetches the list of 1st year students of IUT INFO.